디코딩 결과 값 출력하기 + 측정 데이터 출력

+++ app/src/main/java/io/bitsound/coretestapp/view/MainView.java
@@ -9,6 +9,5 @@ public interface MainView {
     void startPerformanceRecResultActivity(boolean preambleCsSelected,
                                            boolean energyDetectorSelected, boolean qokShapingSelected,
                                            boolean localSyncFinderSelected, int frameType, int coreType, int noSigThreshold,
-                                           int combiningThreshold, double rec, double gamma, double unitBufferSize, int recCount);
-
+                                           int combiningThreshold, double rec, double gamma, int unitBufferSize, int recCount);
 }
